+++ src/linalg/vector_arithmetic.c
@@ -28,7 +28,7 @@ t_vec	t_vec_sub(t_vec a, t_vec b)
 	return (a);
 }
 
-double	t_vec_dot(t_vec a, t_vec b)
+REAL	t_vec_dot(t_vec a, t_vec b)
 {
 	return (a.x * b.x + a.y * b.y + a.z * b.z);
 }
